Migrate components to their own individual files for better organization.
- Add multiple .tsx files for the various home screen components.
- Modify the home page.tsx to instead utilize the new components rather than having it all in the page.tsx file itself.
