+++ src/components/SequenceIconBar.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+function SequenceIconBar() {
+  return (
+    <div className={"sequence-icon-box"}>
+      <div className={"title-box"}>
+        <span className="title-box-text">SEQUENCE</span>
+      </div>
+    </div>
+  )
+}
+
+export default SequenceIconBar
