+++ src/components/TopBar.tsx
@@ -4,7 +4,6 @@ import {Button} from "@/components/ui/button";
 import React from "react";
 import Image from "next/image";
 import RightArrowIcon from "../../public/svg/RightArrow.svg";
-import LeftArrowIcon from "../../public/svg/LeftArrow.svg";
 import Link from "next/link";
 
 interface TopBarProps {
@@ -44,23 +43,6 @@ const TopBar: React.FC<TopBarProps> = ({ buttonLabels, showMoreButton=false }) =
             </>
           )}
         </Button>
-      </div>
-      <div className={"memory-bar"}>
-        <div className={"left-arrow-box"}>
-          <Image className={"left-arrow-icon"}
-                 priority
-                 src={LeftArrowIcon}
-                 alt={"Left Arrow"}
-          />
-        </div>
-        <div className={"memory-box"}>MEMORY: 1</div>
-        <div className={"right-arrow-box"}>
-          <Image className={"right-arrow-icon"}
-                 priority
-                 src={RightArrowIcon}
-                 alt={"Right Arrow"}
-          />
-        </div>
       </div>
     </div>
   );
