+++ src/components/home/LeftMeterBox.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+function LeftMeterBox () {
+  return (
+    <div className={"left-meter-box"}>
+      <div className={"title-box"}>
+        <span className="title-box-text">VOLTAGE</span>
+      </div>
+    </div>
+  )
+}
+
+export default LeftMeterBox
