+++ src/components/home/MiddleBox.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+function MiddleBox () {
+  return (
+    <div className={"middle-box"}>
+      <div className={"title-box"}>
+        <span className="title-box-text text-nowrap">WELD PARAMETERS</span>
+      </div>
+    </div>
+  )
+}
+
+export default MiddleBox
