+++ src/components/home/RightMeterBox.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+function RightMeterBox () {
+  return (
+    <div className={"right-meter-box"}>
+      <div className={"title-box"}>
+        <span className="title-box-text text-nowrap">WIRE SPEED</span>
+      </div>
+    </div>
+  )
+}
+
+export default RightMeterBox
